broke up index.js into two files
