+++ src/server/index.test.js
@@ -1,5 +1,5 @@
 
-const app = require('./index');
+const app = require('./app');
 const supertest = require('supertest');
 const request = supertest(app);
